[camera] Add Camera Component
Add a new component 'Camera.vue' from Home.vue.
