+++ frontend/src/views/Home.vue
@@ -1,39 +1,18 @@
 <template>
-  <div class="home">
-    <video id="video" autoplay playsinline></video>
-    <input type="file" accept="image/*" capture="camera" />
-    <button>촬영</button>
-  </div>
+  <camera-view></camera-view>
 </template>
 
 <script>
-const userMedia = navigator.mediaDevices.getUserMedia({
-  audio: false,
-  video: {
-    width: 640,
-    height: 480,
-    facinMode: 'environment'
-  }
-})
-userMedia.then(function (stream) {
-  const videoEl = document.getElementById('video')
-  const videoTracks = stream.getVideoTracks()[0]
-  console.log('cccc : ' + videoTracks.label)
-  videoEl.srcObject = stream
-}).catch(function (error) {
-  console.error(error.message)
-})
+import Camera from '../components/Camera.vue'
+
 export default {
   name: 'Home',
   components: {
-  },
-  data () {
-    return {
-      showCamera: false
-    }
+    'camera-view': Camera
   }
 }
 </script>
 
 <style lang="scss" scoped>
+
 </style>
